Allow invoking overloaded jmx methods.
By default the invoke fn will pick the first jmx method signature that
it finds. This is normally fine, but in the case of overloaded jmx
methods may not be correct. This patch keeps the default behavior of the
invoke method but allows for specifying a cusom method signature in the
case that an overloaded method needs to be invoked.